requirements.txt
python-dotenv==1.0.1
pydantic==2.8.2
requests==2.32.3